+++ Sources/Utils/LogUtils/DbgFormatter.cs
@@ -56,18 +56,20 @@ public static class DbgFormatter {
   /// into a human readable string.</remarks>
   /// <param name="collection">A collection to represent as a string.</param>
   /// <param name="predicate">A predicate to use to extract string representation of an item. If
+  /// <param name="separator">String to use to glue the parts.</param>
   /// <c>null</c> then standard <c>ToString()</c> is used.</param>
   /// <returns>Human readable form of the collection.</returns>
   /// <typeparam name="TSource">Collection's item type.</typeparam>
-  public static string C2S<TSource>(
-      IEnumerable<TSource> collection, Func<TSource, string> predicate = null) {
+  public static string C2S<TSource>(IEnumerable<TSource> collection,
+                                    Func<TSource, string> predicate = null,
+                                    string separator = ",") {
     var res = new StringBuilder();
     var firstItem = true;
     foreach (var item in collection) {
       if (firstItem) {
         firstItem = false;
       } else {
-        res.Append(',');
+        res.Append(separator);
       }
       if (predicate != null) {
         res.Append(predicate(item));